Bugfix preventing the reset of the remote device
For many devices that use the DTR line as a reset signal, it is not a good idea to always keep resetting the device on every connection. A new connection (by a new instance of the host program) might be made at any time and therefore the execution state on the connected device should be preserved. On Linux, the device is always reset automatically, if the Hangup flag is not disabled.
I hope this will be integrated soon :-)